+++ manta-pay/src/test/transfer.rs
@@ -103,6 +103,48 @@ fn reclaim() {
     );
 }
 
+/// Tests that `generate_proof_input` from [`Transfer`] and [`TransferPost`] gives the same [`ProofInput`].
+#[test]
+fn generate_proof_input_is_compatibile() {
+    let mut rng = OsRng;
+    assert!(
+        matches!(
+            Mint::sample_and_check_generate_proof_input_compatibility(
+                &(),
+                &rng.gen(),
+                &mut UtxoAccumulator::new(rng.gen()),
+                &mut rng
+            ),
+            Ok(true),
+        ),
+        "For a random Mint, `generate_proof_input` from `Transfer` and `TransferPost` should have given the same `ProofInput`."
+    );
+    assert!(
+        matches!(
+            PrivateTransfer::sample_and_check_generate_proof_input_compatibility(
+                &(),
+                &rng.gen(),
+                &mut UtxoAccumulator::new(rng.gen()),
+                &mut rng
+            ),
+            Ok(true),
+        ),
+        "For a random PrivateTransfer, `generate_proof_input` from `Transfer` and `TransferPost` should have given the same `ProofInput`."
+    );
+    assert!(
+        matches!(
+            Reclaim::sample_and_check_generate_proof_input_compatibility(
+                &(),
+                &rng.gen(),
+                &mut UtxoAccumulator::new(rng.gen()),
+                &mut rng
+            ),
+            Ok(true),
+        ),
+        "For a random Reclaim, `generate_proof_input` from `Transfer` and `TransferPost` should have given the same `ProofInput`."
+    );
+}
+
 /// Asserts that `proof` can be SCALE encoded and decoded with at least [`Vec`], [`Cursor`], and
 /// [`File`](std::fs::File).
 #[inline]